\name{BASE Global Environment}
\alias{.pbdBASEEnv}
\docType{data}
\title{Global Environment for the pbdBASE Package}
\description{
  The environment for the pbdBASE package where "global" variables are
  stored.
}
\details{
  The \code{.__blacs_gridinfo_} and \code{._blacs_initialized} objects
  are stored in this environment.
}
\seealso{InitGrid}
