% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globalVariables.r
\docType{data}
\name{BASE Global Environment}
\alias{BASE Global Environment}
\alias{.pbdBASEEnv}
\title{Global Environment for the pbdBASE Package}
\format{An object of class \code{environment} of length 0.}
\usage{
.pbdBASEEnv
}
\description{
The environment for the pbdBASE package where "global" variables are stored.
}
\details{
The \code{.__blacs_gridinfo_} and \code{._blacs_initialized} objects are
stored in this environment.
}
\keyword{datasets}
