% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{base.blacs_gridinit}
\alias{base.blacs_gridinit}
\title{Creating Grid From A System Context}
\usage{
base.blacs_gridinit(SYSCTXT, NPROW, NPCOL, nprocs = pbdMPI::comm.size(comm),
  comm = .pbd_env$SPMD.CT$comm)
}
\arguments{
\item{SYSCTXT}{System context obtained from a call to `sys2blacs_handle`}

\item{NPROW}{Number of rows in the process grid}

\item{NPCOL}{Number of columns in the process grid}

\item{nprocs}{Number of processors in the communicator}

\item{comm}{An MPI (not BLACS) communicator.}
}
\value{
A blacs context number
}
\description{
Creates a grid from a System Context obtained from a call to `sys2blacs_handle`.
}
