% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plinalg.r
\name{base.crossprod}
\alias{base.crossprod}
\title{crossprod}
\usage{
base.crossprod(uplo, trans, x, descx, descc)
}
\arguments{
\item{uplo}{Triangle whose values to use.}

\item{trans}{tcrossprod or crossprod.}

\item{x}{Matrix to crossprod.}

\item{descx}{ScaLAPACK descriptor array.}

\item{descc}{ScaLAPACK descriptor array of output.}
}
\description{
Crossproduct.
}
\details{
For advanced users only.
}
