% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.dallreduce}
\alias{base.dallreduce}
\title{dallreduce}
\usage{
base.dallreduce(x, descx, op = "sum", scope = "All")
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{op}{Operation.}

\item{scope}{Rows, columns, or both.}
}
\description{
Allreduce
}
\details{
For advanced users only.
}
