% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.det}
\alias{base.det}
\title{det}
\usage{
base.det(a, desca)
}
\arguments{
\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Determinant.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
