% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.dhilbmk}
\alias{base.dhilbmk}
\title{dhilbmk}
\usage{
base.dhilbmk(n)
}
\arguments{
\item{n}{Size.}
}
\description{
Create Hilbert matrix.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
