% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack_utility.r
\name{base.dim0}
\alias{base.dim0}
\title{maxdim}
\usage{
base.dim0(dim, ICTXT = 0)
}
\arguments{
\item{dim}{Global dim.}

\item{ICTXT}{BLACS context.}
}
\description{
Compute dimensions on process MYROW=MYCOL=0
}
\details{
For advanced users only.
}

