% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{base.free_blacs_system_handle}
\alias{base.free_blacs_system_handle}
\title{Free Blacs System Handle}
\usage{
base.free_blacs_system_handle(SHANDLE)
}
\arguments{
\item{SHANDLE}{A system handle. Obtained via a call to `sys2blacs.handle`}
}
\value{
None
}
\description{
Free Blacs System Handle
}
