% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matexp.r
\name{base.matexp}
\alias{base.matexp}
\title{matexp}
\usage{
base.matexp(A, p = 6, t = 1)
}
\arguments{
\item{A}{Matrix to exponentiate.}

\item{p}{Pade' expansion size.}

\item{t}{Scaling factor.}
}
\value{
exp(A)
}
\description{
Serial matrix exponentiation.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
