% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.r
\name{base.maxdim}
\alias{base.maxdim}
\title{maxdim}
\usage{
base.maxdim(dim)
}
\arguments{
\item{dim}{Global dim.}
}
\value{
Maximum dimension.
}
\description{
Compute maximum dimension across all nodes
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
