% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_p_matexp.r
\name{base.p_matexp_pade_wrap}
\alias{base.p_matexp_pade_wrap}
\title{p_matexp_pade_wrap}
\usage{
base.p_matexp_pade_wrap(A, desca, p = 6)
}
\arguments{
\item{A}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}

\item{p}{Order of the Pade' approximation.}
}
\description{
Pade' expansion.
}
\details{
For advanced users only.
}

