% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_variance.r
\name{base.pdclvar}
\alias{base.pdclvar}
\title{Column Variances}
\usage{
base.pdclvar(x, descx)
}
\arguments{
\item{x}{The matrix.}

\item{descx}{ScaLAPACK descriptor array.}
}
\description{
Computes the variances of a ScaLAPCK-like distributed matrix.
Significantly faster than using \code{apply()}, even in compared
to the performance differences you would find comparing these
two approaches using just base R.
}
