% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.pdhilbmk}
\alias{base.pdhilbmk}
\title{pdhilbmk}
\usage{
base.pdhilbmk(descx)
}
\arguments{
\item{descx}{ScaLAPACK descriptor matrix.}
}
\description{
Create Hilbert matrix.
}
\details{
For advanced users only.
}

