% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.pdmkcpn1}
\alias{base.pdmkcpn1}
\title{pdmkcpn1}
\usage{
base.pdmkcpn1(coef, descx)
}
\arguments{
\item{coef}{Coefficients vector.}

\item{descx}{ScaLAPACK descriptor array.}
}
\description{
Create Companion Matrix
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
