% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_prblas.r
\name{base.pdmvsum}
\alias{base.pdmvsum}
\title{R-like Matrix-Vector Sum}
\usage{
base.pdmvsum(x, descx, y, descy)
}
\arguments{
\item{x}{Matrix.}

\item{descx, descy}{ScaLAPACK descriptor array.}

\item{y}{Vector.}
}
\description{
For advanced users only.
}

