% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.pdsweep}
\alias{base.pdsweep}
\title{pdsweep}
\usage{
base.pdsweep(x, descx, vec, MARGIN, FUN)
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{vec}{Vector}

\item{MARGIN}{Rows or columns.}

\item{FUN}{Function.}
}
\description{
Matrix-Vector Sweep
}
\details{
For advanced users only.
}
