% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{base.procgrid}
\alias{base.procgrid}
\title{procgrid}
\usage{
base.procgrid(nprocs)
}
\arguments{
\item{nprocs}{Number of processors.}
}
\description{
"Optimal" process grid when nprow and npcol are empty
}
\details{
For advanced users only.
}

