% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_prblas.r
\name{base.rcolcpy}
\alias{base.rcolcpy}
\title{R Column Copy}
\usage{
base.rcolcpy(x, descx, y, descy, xcol, ycol)
}
\arguments{
\item{x, y}{Matrix.}

\item{descx, descy}{ScaLAPACK descriptor array.}

\item{xcol, ycol}{Columns.}
}
\description{
For advanced users only. See pbdDMAT for high-level functions.
}
