% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_prblas.r
\name{base.rl2blas}
\alias{base.rl2blas}
\title{Level 2 R-like BLAS}
\usage{
base.rl2blas(x, descx, vec, FUN)
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{vec}{Global vector.}

\item{FUN}{Function.}
}
\description{
For advanced users only.
}
