% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_prblas.r
\name{base.rl2insert}
\alias{base.rl2insert}
\title{R-like Matrix-Vector Insertion}
\usage{
base.rl2insert(x, descx, vec, i, j)
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{vec}{Global vector.}

\item{i, j}{Indices.}
}
\description{
For advanced users only.
}

