% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdgecon}
\alias{base.rpdgecon}
\title{rpdgecon}
\usage{
base.rpdgecon(norm, m, n, a, desca)
}
\arguments{
\item{norm}{Type of norm.}

\item{m, n}{Problem size}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Inverse condition number of a general matrix.
}
\details{
For advanced users only.
}

