% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_qr.r
\name{base.rpdgelqf}
\alias{base.rpdgelqf}
\title{rpdgelqf}
\usage{
base.rpdgelqf(m, n, x, descx)
}
\arguments{
\item{m, n}{Problem size.}

\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}
}
\description{
LQ.
}
\details{
For advanced users only.
}
