% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_lm.R
\name{base.rpdgels}
\alias{base.rpdgels}
\title{rpdgels}
\usage{
base.rpdgels(tol, m, n, nrhs, a, desca, b, descb)
}
\arguments{
\item{tol}{Numerical tolerance for the QR.}

\item{m, n}{Problem size.}

\item{nrhs}{Number of right hand sides.}

\item{a}{Left hand side.}

\item{desca}{ScaLAPACK descriptor array.}

\item{b}{Right hand side.}

\item{descb}{ScaLAPACK descriptor array.}
}
\description{
Linear model fitter via rank-revealing QR (with pivoting).
}
\details{
For advanced users only.
}

