% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdgemr2d}
\alias{base.rpdgemr2d}
\title{rpdgemr2d}
\usage{
base.rpdgemr2d(x, descx, descy)
}
\arguments{
\item{x}{Matrix.}

\item{descx, descy}{ScaLAPACK descriptor array.}
}
\description{
General 2d block cyclic redistribution function.
}
\details{
For advanced users only.
}
