% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_qr.r
\name{base.rpdgeqpf}
\alias{base.rpdgeqpf}
\title{rpdgeqpf}
\usage{
base.rpdgeqpf(tol, m, n, x, descx)
}
\arguments{
\item{tol}{Numerical tolerance for the QR.}

\item{m,n}{Problem size.}

\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{nrhs}{Number of right hand sides.}
}
\description{
QR.
}
\details{
For advanced users only.
}

