% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdgesv}
\alias{base.rpdgesv}
\title{rpdgesv}
\usage{
base.rpdgesv(n, nrhs, a, desca, b, descb)
}
\arguments{
\item{n}{Problem size.}

\item{nrhs}{Number of right hand sides.}

\item{a, b}{Matrix.}

\item{desca, descb}{ScaLAPACK descriptor array.}
}
\description{
Solving a (square) system of equations.
}
\details{
For advanced users only.
}

