% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdgesvd}
\alias{base.rpdgesvd}
\title{rpdgesvd}
\usage{
base.rpdgesvd(
  jobu,
  jobvt,
  m,
  n,
  a,
  desca,
  descu,
  descvt,
  ...,
  inplace = FALSE,
  comm = .pbd_env$SPMD.CT$comm
)
}
\arguments{
\item{jobu, jobvt}{Control for u/vt return.}

\item{m, n}{Problem size.}

\item{a}{Matrix.}

\item{desca, descu, descvt}{ScaLAPACK descriptor array.}

\item{...}{Ignored}

\item{inplace}{Should the computation be done in-place or not.  For REALLY advanced users only.}

\item{comm}{An MPI (not BLACS) communicator.}
}
\description{
SVD.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
