% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdgetrf}
\alias{base.rpdgetrf}
\title{rpdgetrf}
\usage{
base.rpdgetrf(a, desca)
}
\arguments{
\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
LU factorization.
}
\details{
For advanced users only.
}
