% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdgetri}
\alias{base.rpdgetri}
\title{rpdgetri}
\usage{
base.rpdgetri(n, a, desca)
}
\arguments{
\item{n}{Problem size.}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Matrix inversion.
}
\details{
For advanced users only.
}
