% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdlange}
\alias{base.rpdlange}
\title{rpdlange}
\usage{
base.rpdlange(norm, m, n, a, desca)
}
\arguments{
\item{norm}{Type of norm.}

\item{m, n}{Problem size}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Matrix norms.
}
\details{
For advanced users only.
}

