% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{base.rpdlaprnt}
\alias{base.rpdlaprnt}
\title{rpdlaprnt}
\usage{
base.rpdlaprnt(m, n, a, desca)
}
\arguments{
\item{m, n}{Number rows/cols.}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Matrix printer.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
