% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.r
\name{base.rpdorgqr}
\alias{base.rpdorgqr}
\title{rpdorgqr}
\usage{
base.rpdorgqr(m, n, k, qr, descqr, tau)
}
\arguments{
\item{m, n}{Problem size.}

\item{k}{Number of elementary reflectors.}

\item{qr}{QR decomposition.}

\item{descqr}{ScaLAPACK descriptor array.}

\item{tau}{Elementary reflectors.}
}
\value{
Q matrix of the QR decomposition.
}
\description{
Recover Q.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
