% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_qr.r
\name{base.rpdormqr}
\alias{base.rpdormqr}
\title{rpdormqr}
\usage{
base.rpdormqr(side, trans, m, n, k, qr, descqr, tau, c, descc)
}
\arguments{
\item{side}{}

\item{trans}{Q or Q^T.}

\item{m,n}{Problem size.}

\item{k}{Number of elementary reflectors.}

\item{qr}{QR decomposition.}

\item{descqr}{ScaLAPACK descriptor array.}

\item{tau}{Elementary reflectors.}

\item{c}{Vector.}

\item{descc}{ScaLAPACK descriptor array.}
}
\description{
Qy.
}
\details{
For advanced users only.
}

