% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdpotrf}
\alias{base.rpdpotrf}
\title{rpdpotrf}
\usage{
base.rpdpotrf(uplo, n, a, desca)
}
\arguments{
\item{uplo}{Triangle where the information is stored (in the symmetric matrix).}

\item{n}{Problem size.}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Cholesky factorization.
}
\details{
For advanced users only.
}
