% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdsyev}
\alias{base.rpdsyev}
\title{rpdsyev}
\usage{
base.rpdsyev(jobz, uplo, n, a, desca, descz)
}
\arguments{
\item{jobz}{Control for if vectors/values/both are returned.}

\item{uplo}{Triangle where the information is stored (in the symmetric matrix).}

\item{n}{Problem size.}

\item{a}{Matrix.}

\item{desca, descz}{ScaLAPACK descriptor array.}
}
\description{
Symmetric eigenvalue decomposition.
}
\details{
For advanced users only.
}

