% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdsyevx}
\alias{base.rpdsyevx}
\title{rpdsyevx}
\usage{
base.rpdsyevx(jobz, range, n, a, desca, vl, vu, il, iu, abstol = 1e-08,
  orfac = 0.001)
}
\arguments{
\item{jobz}{Control for if vectors/values/both are returned.}

\item{range}{Parameter to determine the search criteria for eigenvalues.}

\item{n}{Problem size.}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}

\item{vl, vu}{Endpoints of the interval subset of the real line in which to search for eigenvalues, if specified by \code{range}.}

\item{il, iu}{Eigenvalues with indices \code{il}, ..., \code{iu} will be found, if specified by \code{range}.}

\item{abstol}{Absolute error tolerance for the eigenvalues.}

\item{orfac}{Eigenvectors with eigenvalues below orfac*norm(a) of each other are reorthogonalized.}
}
\description{
Genearlized eigenvalue problem.
}
\details{
For advanced users only.
}
