% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pblas_level_3.r
\name{base.rpdtran}
\alias{base.rpdtran}
\title{rpdtran}
\usage{
base.rpdtran(a, desca, descc)
}
\arguments{
\item{a}{Matrix.}

\item{desca,descc}{ScaLAPACK descriptor array.}
}
\description{
Transpose.
}
\details{
For advanced users only.
}

