% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack.r
\name{base.rpdtrcon}
\alias{base.rpdtrcon}
\title{rpdtrcon}
\usage{
base.rpdtrcon(norm, uplo, diag, n, a, desca)
}
\arguments{
\item{norm}{Type of norm.}

\item{uplo}{Triangle where information is stored.}

\item{diag}{Specifies if the matrix is unit triangular or not.}

\item{n}{Problem size}

\item{a}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}
}
\description{
Inverse condition number of a triangular matrix.
}
\details{
For advanced users only.
}

