% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_prblas.r
\name{base.rrowcpy}
\alias{base.rrowcpy}
\title{R Row Copy}
\usage{
base.rrowcpy(x, descx, y, descy, xrow, yrow)
}
\arguments{
\item{x, y}{Matrix.}

\item{descx, descy}{ScaLAPACK descriptor array.}

\item{xrow, yrow}{Rows.}
}
\description{
For advanced users only. See pbdDMAT for high-level functions.
}
