% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.tri2zero}
\alias{base.tri2zero}
\title{tri2zero}
\usage{
base.tri2zero(x, descx, uplo = "L", diag = "N")
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{uplo}{Triangle.}

\item{diag}{Zero diagonal as well.}
}
\description{
Zero Triangle
}
\details{
For advanced users only.
}

