% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_blacs.r
\name{base.valid_context}
\alias{base.valid_context}
\title{BLACS Context Validation}
\usage{
base.valid_context(ICTXT, ..., override = FALSE)
}
\arguments{
\item{ICTXT}{BLACS context number.}

\item{...}{Not used.}

\item{override}{If \code{override=FALSE}, the context number will produce an
error if it is any of the reserved contexts (0, 1, or 2).}
}
\description{
Checks if a supplied \code{ICTXT} is valid.
}
