% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs_reductions.r
\name{base.igamx2d}
\alias{base.igamx2d}
\title{BLACS Max}
\usage{
base.igamx2d(ICTXT, SCOPE, m, n, x, lda, RDEST, CDEST)
}
\arguments{
\item{ICTXT}{BLACS ICTXT.}

\item{SCOPE}{Rows, cols, or both.}

\item{m, n}{Problem size.}

\item{x}{Local values.}

\item{lda}{Leading dimension.}

\item{RDEST}{Row destination.}

\item{CDEST}{Col destination.}
}
\description{
Max value across a process grid.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
