% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs_reductions.r
\name{base.dgamx2d}
\alias{base.dgamn2d}
\alias{base.dgamx2d}
\alias{base.igamn2d}
\title{BLACS Min}
\usage{
base.dgamx2d(ICTXT, SCOPE, m, n, x, lda, RDEST, CDEST)

base.igamn2d(ICTXT, SCOPE, m, n, x, lda, RDEST, CDEST)

base.dgamn2d(ICTXT, SCOPE, m, n, x, lda, RDEST, CDEST)
}
\arguments{
\item{ICTXT}{BLACS ICTXT.}

\item{SCOPE}{Rows, cols, or both.}

\item{m, n}{Problem size.}

\item{x}{Local values.}

\item{lda}{Leading dimension.}

\item{RDEST}{Row destination.}

\item{CDEST}{Col destination.}
}
\description{
Min value across a process grid.
}
\details{
For advanced users only.
}

