% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs_reductions.r
\name{base.dgesd2d}
\alias{base.dgerv2d}
\alias{base.dgesd2d}
\title{BLACS Point to Poin}
\usage{
base.dgesd2d(ICTXT, SCOPE, m, n, x, lda, RDEST, CDEST)

base.dgerv2d(ICTXT, SCOPE, m, n, x, lda, RDEST, CDEST)
}
\arguments{
\item{ICTXT}{BLACS ICTXT.}

\item{SCOPE}{Rows, cols, or both.}

\item{m, n}{Problem size.}

\item{x}{Local values.}

\item{lda}{Leading dimension.}

\item{RDEST}{Row destination.}

\item{CDEST}{Col destination.}
}
\description{
Sent value across a process grid.
}
\details{
For advanced users only.
}

