% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{coords}
\alias{coords}
\alias{indxg2l}
\alias{indxl2g}
\title{Local to Global/Global to Local Indexing}
\usage{
indxg2l(INDXGLOB, NB, IPROC, ISRCPROC, NPROCS)

indxl2g(INDXLOC, NB, IPROC, ISRCPROC, NPROCS)
}
\arguments{
\item{INDXGLOB}{Global index.}

\item{NB}{Block size.}

\item{IPROC}{Coordinate of the process whose local info is to be determined.}

\item{ISRCPROC}{The coordinate of the process that possesses the first row/column of the distributed matrix.  That's always 0 pbdDMAT.}

\item{NPROCS}{Total number of processors over which matrix is distributed.}

\item{INDXLOC}{Local index.}
}
\description{
Get the local index given global information.
}
\details{
For advanced users only.
}

