% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coords.r
\name{coordspair}
\alias{coordspair}
\alias{g2lpair}
\alias{l2gpair}
\title{g2lpair}
\usage{
g2lpair(gi, gj, bldim, ICTXT)

l2gpair(i, j, bldim, ICTXT)
}
\arguments{
\item{gi,gj}{Global indices.}

\item{bldim}{Blocking dimension}

\item{ICTXT}{BLACS context.}

\item{i,j}{Local indices.}
}
\description{
Get the local index-pair given global information.
}
\details{
For advanced users only.
}

