% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack_utility.r
\name{g2l_coord}
\alias{base.g2l_coord}
\alias{g2l_coord}
\title{g2l_coord}
\usage{
base.g2l_coord(ind, dim, bldim, ICTXT = 0)

g2l_coord(ind, dim, bldim, ICTXT = 0)
}
\arguments{
\item{ind}{Matrix indices.}

\item{dim}{Global dim.}

\item{bldim}{Blocking dimension.}

\item{ICTXT}{BLACS context.}
}
\description{
Global to local coords.
}
\details{
For advanced users only.
}

