% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{g2lcoord}
\alias{g2lcoord}
\title{g2lcoord}
\usage{
g2lcoord(dim, bldim, gi, gj, gridinfo)
}
\arguments{
\item{dim}{Global dimension.}

\item{bldim}{Blocking dimension.}

\item{gi, gj}{Global row and column indices, respectively.}

\item{gridinfo}{The return of \code{base.blacs(ICTXT(x))}.  See the Details section
for more information.}
}
\value{
For the process that owns the desired local data at global indices
\code{(gi, gj)}, the return is the local index.  Otherwise, \code{NA}
is returned.
}
\description{
Global to local coordinates with explicit ownership given.
}
