% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{get.comm.from.ICTXT}
\alias{get.comm.from.ICTXT}
\title{Getting Communicator From BLACS Context}
\usage{
get.comm.from.ICTXT(ICTXT)
}
\arguments{
\item{ICTXT}{a BLACS context}
}
\description{
Blacs context are associated with a certain communicator. It can be useful to retrieve this communicator to manipulate the matrix accordingly.
}
