% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{gridexit}
\alias{gridexit}
\alias{base.gridexit}
\alias{gridexit}
\title{gridexit}
\usage{
base.gridexit(ICTXT, override = FALSE)

gridexit(ICTXT, override = FALSE)
}
\arguments{
\item{ICTXT}{BLACS context number.}

\item{override}{logical; if TRUE, ignores normal check preventing the
closing of \code{ICTXT} values of 0, 1, and 2. This could cause things
to go crazy and I do not recommend it.}
}
\value{
Silently returns 0 when successful. Silently returns 1 when
requested \code{ICTXT} does not exist.
}
\description{
Frees a BLACS context.
}
\details{
For advanced users only.

The function frees the requested BLACS context. It is a trivial wrapper for
the BLACS routine \code{BLACS_GRIDEXIT}. Also removes the object
\code{.__blacs_gridinfo_ICTXT}.

Contexts 0, 1, and 2 can not be freed in this way unless the argument
\code{override=FALSE}. This will probably break something and I do not
recommend it.
}
\keyword{BLACS}
