% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack_utility.r
\name{l2g_coord}
\alias{l2g_coord}
\alias{base.l2g_coord}
\alias{l2g_coord}
\title{l2g_coord}
\usage{
base.l2g_coord(ind, bldim, ICTXT = 0, dim = NULL)

l2g_coord(ind, bldim, ICTXT = 0, dim = NULL)
}
\arguments{
\item{ind}{Matrix indices.}

\item{bldim}{Blocking dimension.}

\item{ICTXT}{BLACS context.}

\item{dim}{Ignored; will be removed in a future version.}
}
\description{
Local to global coords.
}
\details{
For advanced users only.
}
