% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{numroc2}
\alias{numroc2}
\title{numroc2}
\usage{
numroc2(N, NB, IPROC, NPROCS)
}
\arguments{
\item{N}{Global number of rows/cols.}

\item{NB}{Block size.}

\item{IPROC}{Coordinate of the process whose local info is to be determined.}

\item{NPROCS}{Total number of processors over which matrix is distributed.}
}
\description{
A better version of NUMROC (NUMber Rows Or Columns).  Returns the local
dimension given global matrix + distribution parameters.
}
\details{
For advanced users only.
}
