% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdBASE-package.r
\docType{package}
\name{pbdBASE-package}
\alias{pbdBASE-package}
\title{ScaLAPACK Wrappers and Utilities}
\description{
A package contains the basic methods for dealing with distributed data
types, as well as the data types themselves.
}
\details{
\tabular{ll}{ Package: \tab pbdBASE\cr Type: \tab Package\cr License: \tab
MPL\cr LazyLoad: \tab yes\cr } This package requires an MPI library
(OpenMPI, MPICH2, or LAM/MPI).
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}, Wei-Chen Chen, George
Ostrouchov, and Pragneshkumar Patel.
}
\keyword{Package}
