% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{sys2blacs.handle}
\alias{sys2blacs.handle}
\title{Context Within a Given Communicator}
\usage{
sys2blacs.handle(comm)
}
\arguments{
\item{comm}{Communicator for which you want to set the BLACS context}
}
\value{
A system handle, i.e. the system context number. System contexts can be used to have ScalaPACK methods run in different communicators.
}
\description{
Creates a context that will be valid for a given communicator
}
\seealso{
base.free_blacs_system_handle, base.blacs_gridinit
}
