\name{base.internal}

\alias{base.blacs}
\alias{base.blacsexit}
\alias{base.crossprod}
\alias{base.dallreduce}
\alias{base.ddiagmk}
\alias{base.ddiagtk}
\alias{base.descinit}
\alias{base.gridexit}
\alias{base.indxg2p}
\alias{base.init.grid}
\alias{base.minctxt}
\alias{base.mkgblmat}
\alias{base.mksubmat}
\alias{base.numroc}
\alias{base.ownany}
\alias{base.pcoord}
\alias{base.pdchtri}
\alias{base.pdsweep}
\alias{base.pnum}
\alias{base.procgrid}
\alias{base.rcolcpy}
\alias{base.rl2blas}
\alias{base.rl2insert}
\alias{base.rpdgecon}
\alias{base.rpdgels}
\alias{base.rpdgemm}
\alias{base.rpdgeqpf}
\alias{base.rpdgesv}
\alias{base.rpdgesvd}
\alias{base.rpdgetrf}
\alias{base.rpdgetri}
\alias{base.rpdlange}
\alias{base.rpdlaprnt}
\alias{base.rpdorgqr}
\alias{base.rpdormqr}
\alias{base.rpdpotrf}
\alias{base.rpdtran}
\alias{base.rpdtrcon}
\alias{base.rrowcpy}
\alias{base.tri2zero}
\alias{base.valid_context}


\alias{.conflicts.OK}

\title{ All pbdPASE Internal Functions and Data }
\description{ All pbdBASE internal functions }
\keyword{internal}

