% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack_utility.r
\name{base.descinit}
\alias{base.descinit}
\title{descinit}
\usage{
base.descinit(dim, bldim, ldim, ICTXT = 0)
}
\arguments{
\item{dim}{Global dim.}

\item{bldim}{Blocking dim.}

\item{ldim}{Local dim.}

\item{ICTXT}{BLACS context.}
}
\description{
Creates ScaLAPACK descriptor array.
}
\details{
For advanced users only.
}

