% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack_utility.r
\name{base.maxdim}
\alias{base.maxdim}
\title{maxdim}
\usage{
base.maxdim(dim)
}
\arguments{
\item{dim}{Global dim.}
}
\description{
Compute maximum dimension across all nodes
}
\details{
For advanced users only.
}
