% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{base.nbd}
\alias{base.nbd}
\title{Next Best Divisor}
\usage{
base.nbd(n, d)
}
\arguments{
\item{n}{The divident (number divided into).}

\item{d}{The candidate divisor.}
}
\description{
Given integers n and d, with n>d, this function finds the "next
best divisor" of n which is greater than or equal to d.
}
\details{
Suprisingly useful for thinking about processor grid shapes.
}
\examples{
\dontrun{
library(pbdBASE, quiet = TRUE)
base.nbd(100, 10) # 10 divides 100, so 10 is returned
base.nbd(100, 11) # 11 does not, so the "next best" divisor, 20, is returned
}

}

