% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_scalapack_utility.r
\name{base.numroc}
\alias{base.numroc}
\title{numroc}
\usage{
base.numroc(dim, bldim, ICTXT = 0, fixme = TRUE)
}
\arguments{
\item{dim}{Global dim.}

\item{bldim}{Blocking dim.}

\item{ICTXT}{BLACS context.}

\item{fixme}{Should ldims be "rounded" to 0 or not.}
}
\description{
NUMber of Rows Or Columns
}
\details{
For advanced users only.
}
