% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_p_matexp.r
\name{base.p_matpow_by_squaring_wrap}
\alias{base.p_matpow_by_squaring_wrap}
\title{p_matpow_by_squaring_wrap}
\usage{
base.p_matpow_by_squaring_wrap(A, desca, b = 1)
}
\arguments{
\item{A}{Matrix.}

\item{desca}{ScaLAPACK descriptor array.}

\item{b}{Power.}
}
\description{
Matrix power by squaring.
}
\details{
For advanced users only.
}
