% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plinalg.r
\name{base.pdchtri}
\alias{base.pdchtri}
\title{pdchtri}
\usage{
base.pdchtri(uplo, x, descx, descc)
}
\arguments{
\item{uplo}{Triangle whose values to use.}

\item{x}{Matrix to crossprod.}

\item{descx}{ScaLAPACK descriptor array.}

\item{descc}{ScaLAPACK descriptor array of output.}
}
\description{
Inverse of cholesky.
}
\details{
For advanced users only.
}
