% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist.r
\name{base.redist}
\alias{base.redist}
\title{base.redist}
\usage{
base.redist(desc, A)
}
\arguments{
\item{desc}{ScaLAPACK descriptor array.}

\item{A}{Matrix.}
}
\description{
Redistribute a matrix from rank 0 to all ranks in block cyclic
fashion.
}

