% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.r
\name{base.rpdgeqpf}
\alias{base.rpdgeqpf}
\title{rpdgeqpf}
\usage{
base.rpdgeqpf(tol, m, n, x, descx, comm = .pbd_env$SPMD.CT$comm)
}
\arguments{
\item{tol}{Numerical tolerance for the QR.}

\item{m, n}{Problem size.}

\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{comm}{An MPI (not BLACS) communicator.}
}
\value{
A list contains QR results.
}
\description{
QR.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
