% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.r
\name{base.rpdorglq}
\alias{base.rpdorglq}
\title{rpdorglq}
\usage{
base.rpdorglq(m, n, k, lq, desc, tau)
}
\arguments{
\item{m, n}{Problem size.}

\item{k}{Number of elementary reflectors.}

\item{lq}{QR decomposition.}

\item{desc}{ScaLAPACK descriptor array.}

\item{tau}{Elementary reflectors.}
}
\value{
Q matrix of the QR decomposition.
}
\description{
Recover Q.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
