% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pblas_level3.r
\name{base.rpdtran}
\alias{base.rpdtran}
\title{rpdtran}
\usage{
base.rpdtran(a, desca, descc)
}
\arguments{
\item{a}{Matrix.}

\item{desca, descc}{ScaLAPACK descriptor array.}
}
\description{
Transpose.
}
\details{
For advanced users only.
}

