% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_prblas.r
\name{base.rrowcpy2}
\alias{base.rrowcpy2}
\title{R Row Copy-2}
\usage{
base.rrowcpy2(x, descx, y, descy, xrow, yrow)
}
\arguments{
\item{x, y}{Matrix.}

\item{descx, descy}{ScaLAPACK descriptor array.}

\item{xrow, yrow}{Rows.}
}
\description{
For advanced users only.
}
