% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{diag}
\alias{base.ddiagmk}
\alias{base.ddiagtk}
\alias{diag}
\title{diag}
\usage{
base.ddiagtk(x, descx, proc.dest = "all")

base.ddiagmk(diag, descx)
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{proc.dest}{Who owns the result.}

\item{diag}{Diagonal.}
}
\description{
Grab diagonal or create distributed diagonal matrix.
}
\details{
For advanced users only.
}

