% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{finalizer}
\alias{finalizer}
\alias{base.finalize}
\alias{finalize}
\title{Finalizer}
\usage{
base.finalize(mpi.finalize = .pbd_env$SPMD.CT$mpi.finalize)

finalize(mpi.finalize = .pbd_env$SPMD.CT$mpi.finalize)
}
\arguments{
\item{mpi.finalize}{If MPI should be shut down.}
}
\description{
A replacement for \code{pbdMPI::finalize()} that automatically
shuts BLACS communicators down.
}
