% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.r
\name{g2l_coord}
\alias{g2l_coord}
\alias{base.g2l_coord}
\title{g2l_coord}
\usage{
base.g2l_coord(ind, bldim, ICTXT = 0, dim = NULL)

g2l_coord(ind, bldim, ICTXT = 0, dim = NULL)
}
\arguments{
\item{ind}{Matrix indices.}

\item{bldim}{Blocking dimension.}

\item{ICTXT}{BLACS context.}

\item{dim}{Ignored; will be removed in a future version.}
}
\value{
Local coords.
}
\description{
Global to local coords.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
