% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mpi_blacs.r
\name{gridinit}
\alias{gridinit}
\alias{base.blacs_init}
\alias{blacs_init}
\alias{blacs_gridinit}
\title{blacs_init}
\usage{
base.blacs_init(ICTXT, NPROW, NPCOL, ..., quiet = FALSE)

blacs_init(ICTXT, NPROW, NPCOL, ..., quiet = FALSE)

blacs_gridinit(ICTXT, NPROW, NPCOL, ..., quiet = FALSE)
}
\arguments{
\item{ICTXT}{BLACS context.}

\item{NPROW, NPCOL}{Number of process rows/cols.}

\item{...}{Additional arguments.}

\item{quiet}{Verbose initialization or not.}
}
\value{
None
}
\description{
BLACS grid initialization.
}
\details{
For advanced users only. See pbdDMAT for high-level functions.
}
