% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_putil.R
\name{base.mksubmat}
\alias{base.mkgblmat}
\alias{base.mksubmat}
\title{(Un)Distribute}
\usage{
base.mksubmat(x, descx)

base.mkgblmat(x, descx, rsrc, csrc)
}
\arguments{
\item{x}{Matrix.}

\item{descx}{ScaLAPACK descriptor array.}

\item{rsrc, csrc}{Row/column source.}
}
\description{
(Un)Distribute matrix.
}
\details{
For advanced users only.
}

