\name{base.internal}

\alias{base.as.ddmatrix}
\alias{base.as.matrix}
\alias{base.blacs}
\alias{base.blacsexit}
\alias{base.blacs.max}
\alias{base.blacs.sum}
\alias{base.bldim}
\alias{base.checkem}
\alias{base.ctxt}
\alias{base.descinit}
\alias{base.dim0}
\alias{base.distribute}
\alias{base.dropper}
\alias{base.firstfew}
\alias{base.g2l_coord}
\alias{base.gmat}
\alias{base.gridexit}
\alias{base.init.grid}
\alias{base.insert}
\alias{base.is.ddmatrix}
\alias{base.l2g_coord}
\alias{base.ldim}
\alias{base.low2zero}
\alias{base.mat.to.ddmat}
\alias{base.maxdim}
\alias{base.minctxt}
\alias{base.numroc}
\alias{base.ownany}
\alias{base.pcoord}
\alias{base.rpdgemm}
\alias{base.rpdlaprnt}
\alias{base.rpdpotrf}
\alias{base.rpdtran}
\alias{base.pnum}
\alias{base.procgrid}
\alias{base.reblock}
\alias{base.redistribute}
\alias{base.rpdgels}
\alias{base.rpdgesv}
\alias{base.rpdgesvd}
\alias{base.rpdgetrf}
\alias{base.rpdgetri}
\alias{base.submat}
\alias{base.submatrix}
\alias{base.valid_context}
\alias{base.vecops}
\alias{base.whatdo}

\alias{base.rpdorgqr}
\alias{base.rpdormqr}
\alias{base.rpdtrsv}
\alias{base.rpdtzrzf}
\alias{base.rpdgeqpf}
\alias{base.qr.R}
\alias{base.rbind}
\alias{base.cbind}
\alias{base.rbind2}

\alias{.conflicts.OK}

\title{ All pbdPASE Internal Functions and Data }
\description{ All pbdBASE internal functions }
\keyword{internal}

