\name{base.internal}

\alias{base.blacs}
\alias{base.blacsexit}
\alias{base.crossprod}
\alias{base.dallreduce}
\alias{base.ddiagmk}
\alias{base.ddiagtk}
\alias{base.descinit}
\alias{base.dim0}
\alias{base.gridexit}
\alias{base.indxg2p}
\alias{base.init.grid}
\alias{base.maxdim}
\alias{base.minctxt}
\alias{base.mkgblmat}
\alias{base.mksubmat}
\alias{base.numroc}
\alias{base.pcoord}
\alias{base.pdchtri}
\alias{base.pdsweep}
\alias{base.pnum}
\alias{base.procgrid}
\alias{base.rcolcpy}
\alias{base.rl2blas}
\alias{base.rl2insert}
\alias{base.rpdgecon}
\alias{base.rpdgels}
\alias{base.rpdgemm}
\alias{base.rpdgeqpf}
\alias{base.rpdgesv}
\alias{base.rpdgesvd}
\alias{base.rpdgetrf}
\alias{base.rpdgetri}
\alias{base.rpdlange}
\alias{base.rpdlaprnt}
\alias{base.rpdorgqr}
\alias{base.rpdormqr}
\alias{base.rpdpotrf}
\alias{base.rpdtran}
\alias{base.rpdtrcon}
\alias{base.rrowcpy}
\alias{base.tri2zero}
\alias{base.valid_context}


\alias{base.dgamn2d}
\alias{base.dgamx2d}
\alias{base.dgerv2d}
\alias{base.dgesd2d}
\alias{base.dgsum2d}
\alias{base.g2l_coord}
\alias{base.igamn2d}
\alias{base.igamx2d}
\alias{base.igsum2d}
\alias{base.l2g_coord}
\alias{base.pdclvar}
\alias{base.rpdgemr2d}
\alias{base.pdmvsum}

\alias{base.nbd}
\alias{base.blacs_gridinit}

\alias{base.rcolcpy2}
\alias{base.rrowcpy2}

\alias{base.rpdsyev}

\alias{.conflicts.OK}

\title{ All pbdPASE Internal Functions and Data }
\description{ All pbdBASE internal functions }
\keyword{internal}

